Bug fix: only call mListener.playFiles once for a folder playback
The call to mListener.playFiles was erroneously inside the loop that generated
the MediaQueueItem list. This resulted in the receiver device being told to play
queues of size [1, 2, 3, ..., n] - each call of which started a new playback
session with a random start point in the (new) queue. Needless to say, this
fixes a lot of the weird behavior that was observed when playing a folder.

Note to self: this is why one is meant to have unittests