Add padding to the bottom of the fragment ListViews
This is somewhat of a hack to make sure that the Cast MiniController doesn't
overlap any important content, but it also provides a nice clear distinction
that you are at the end of the list. And hey, you can now see all the
files/folders in the list, so that's good.