Add basic bookmark support for UPnP containers
Introduces the ability for users to bookmark container items, which then show up
in a separate section in the server browser list (only when the relevant device
is added). When a bookmarked container is clicked on the user is taken there
directly. Pressing 'back' on a bookmarked container returns the user to the
server browser, because we have no way of knowing the parent chain for that
container.